pyproject.toml
[tool.poetry]
name = "dungeonmasterbot"
version = "0.1.0"
description = ""
authors = ["Robert Nasuti <rlnasuti@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
openai = "^0.28.1"
tenacity = "^8.2.3"
langchain = "^0.0.308"
tiktoken = "^0.5.1"
pypdf2 = "^3.0.1"
faiss-cpu = "^1.7.4"
flask-cors = "^4.0.0"
chainlit = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
